requirements.txt
livekit-agents 
livekit-plugins-openai 
livekit-plugins-silero 
livekit-plugins-deepgram
python-dotenv
cerebras_cloud_sdk